refactor(noodles): rename kawaii-pride key + folder to transgender-visibility-day
Keeps key/slug/folder/import name consistent now that the noodle is
named Transgender Visibility Day. URL is /noodles/transgender-visibility-day,
SFC lives at Noodle/TransgenderVisibility/Logo.vue, lookup id is
transgender-visibility-day.
